# packages/server/src/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "Boardie",
    "description": "API to serve the Boardie application"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/matches/:matchId/players": {
      "get": {
        "summary": "Gets all players by match id",
        "tags": [
          "matches/players"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "matchId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The players",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "friend": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "ANON_FRIEND",
                              "USER"
                            ]
                          },
                          "fullName": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "type",
                          "fullName"
                        ],
                        "additionalProperties": false,
                        "description": "Generic friend"
                      },
                      "score": {
                        "type": "number",
                        "nullable": true
                      },
                      "isWinner": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "id",
                      "friend",
                      "score",
                      "isWinner"
                    ],
                    "additionalProperties": false,
                    "description": "Data that represents an existent player"
                  }
                }
              }
            }
          },
          "404": {
            "description": "The match was not found"
          }
        }
      },
      "post": {
        "summary": "Creates a player",
        "tags": [
          "matches/players"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "matchId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "friend": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "ANON_FRIEND",
                          "USER"
                        ]
                      }
                    },
                    "required": [
                      "id",
                      "type"
                    ],
                    "additionalProperties": false,
                    "description": "Generic friend identification data"
                  },
                  "score": {
                    "type": "number",
                    "nullable": true
                  },
                  "isWinner": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "friend",
                  "score",
                  "isWinner"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created player",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "friend": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "type": {
                          "type": "string",
                          "enum": [
                            "ANON_FRIEND",
                            "USER"
                          ]
                        },
                        "fullName": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "type",
                        "fullName"
                      ],
                      "additionalProperties": false,
                      "description": "Generic friend"
                    },
                    "score": {
                      "type": "number",
                      "nullable": true
                    },
                    "isWinner": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "friend",
                    "score",
                    "isWinner"
                  ],
                  "additionalProperties": false,
                  "description": "Data that represents an existent player"
                }
              }
            }
          },
          "403": {
            "description": "The user doesn't have permissions to update the match, or doesn't have a friendship with the friend related to the player trying to be created"
          },
          "404": {
            "description": "The match was not found"
          }
        }
      }
    },
    "/matches/:matchId/players/:playerId": {
      "put": {
        "summary": "Updates a player",
        "tags": [
          "matches/players"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "matchId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "playerId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "friend": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "ANON_FRIEND",
                          "USER"
                        ]
                      }
                    },
                    "required": [
                      "id",
                      "type"
                    ],
                    "additionalProperties": false,
                    "description": "Generic friend identification data"
                  },
                  "score": {
                    "type": "number",
                    "nullable": true
                  },
                  "isWinner": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "friend",
                  "score",
                  "isWinner"
                ],
                "additionalProperties": false,
                "description": "Data used to update a player"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated player",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "friend": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "type": {
                          "type": "string",
                          "enum": [
                            "ANON_FRIEND",
                            "USER"
                          ]
                        },
                        "fullName": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "type",
                        "fullName"
                      ],
                      "additionalProperties": false,
                      "description": "Generic friend"
                    },
                    "score": {
                      "type": "number",
                      "nullable": true
                    },
                    "isWinner": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "friend",
                    "score",
                    "isWinner"
                  ],
                  "additionalProperties": false,
                  "description": "Data that represents an existent player"
                }
              }
            }
          },
          "403": {
            "description": "The user doesn't have permissions to update the match, or doesn't have a friendship with the friend related to the player trying to be updated"
          },
          "404": {
            "description": "The match or player was not found"
          }
        }
      },
      "delete": {
        "summary": "Deletes a player by id",
        "tags": [
          "matches/players"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "matchId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "playerId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The player was deleted"
          },
          "403": {
            "description": "The user doesn't have permissions to update the match"
          },
          "404": {
            "description": "The match or player was not found"
          }
        }
      }
    },
...
  }
}